requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# AI and Computer Vision
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.26.0
scikit-image>=0.21.0

# GIS Libraries
gdal>=3.7.0
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
folium>=0.15.0

# Data Processing
pandas>=2.1.0
geojson>=3.1.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.5.0
